saving as draft with empty inputs
